Clean up

+++ src/main/java/com/mjr/extraplanets/planets/Jupiter/CloudRenderJupiter.java
@@ -20,17 +20,17 @@ public class CloudRenderJupiter extends IRenderHandler {
 	public void render(float partialTicks, WorldClient world, Minecraft mc) {
 		GlStateManager.disableCull();
 		Entity entity = mc.getRenderViewEntity();
-		float f = (float) (entity.lastTickPosY + (entity.posY - entity.lastTickPosY) * (double) partialTicks);
+		float f = (float) (entity.lastTickPosY + (entity.posY - entity.lastTickPosY) * partialTicks);
 		Tessellator tessellator = Tessellator.getInstance();
 		WorldRenderer worldrenderer = tessellator.getWorldRenderer();
-		double d0 = (double) ((float) CloudRenderJupiter.cloudTickCounter + partialTicks);
-		double d1 = (entity.prevPosX + (entity.posX - entity.prevPosX) * (double) partialTicks + d0 * 0.029999999329447746D) / 12.0D;
-		double d2 = (entity.prevPosZ + (entity.posZ - entity.prevPosZ) * (double) partialTicks) / 12.0D + 0.33000001311302185D;
+		double d0 = CloudRenderJupiter.cloudTickCounter + partialTicks;
+		double d1 = (entity.prevPosX + (entity.posX - entity.prevPosX) * partialTicks + d0 * 0.029999999329447746D) / 12.0D;
+		double d2 = (entity.prevPosZ + (entity.posZ - entity.prevPosZ) * partialTicks) / 12.0D + 0.33000001311302185D;
 		float f3 = world.provider.getCloudHeight() - 25 - f + 0.33F;
 		int i = MathHelper.floor_double(d1 / 2048.0D);
 		int j = MathHelper.floor_double(d2 / 2048.0D);
-		d1 = d1 - (double) (i * 2048);
-		d2 = d2 - (double) (j * 2048);
+		d1 = d1 - i * 2048;
+		d2 = d2 - j * 2048;
 		mc.renderEngine.bindTexture(CLOUDS_TEXTURES);
 		GlStateManager.enableBlend();
 		GlStateManager.tryBlendFuncSeparate(770, 771, 1, 0);
@@ -48,10 +48,10 @@ public class CloudRenderJupiter extends IRenderHandler {
 		float f13 = f4 * 0.8F;
 		float f14 = f5 * 0.8F;
 		float f15 = f6 * 0.8F;
-		float f17 = (float) MathHelper.floor_double(d1) * 0.00390625F;
-		float f18 = (float) MathHelper.floor_double(d2) * 0.00390625F;
-		float f19 = (float) (d1 - (double) MathHelper.floor_double(d1));
-		float f20 = (float) (d2 - (double) MathHelper.floor_double(d2));
+		float f17 = MathHelper.floor_double(d1) * 0.00390625F;
+		float f18 = MathHelper.floor_double(d2) * 0.00390625F;
+		float f19 = (float) (d1 - MathHelper.floor_double(d1));
+		float f20 = (float) (d2 - MathHelper.floor_double(d2));
 		GlStateManager.scale(12.0F, 1.0F, 12.0F);
 
 		for (int i1 = 0; i1 < 2; ++i1) {
@@ -73,81 +73,81 @@ public class CloudRenderJupiter extends IRenderHandler {
 			for (int j1 = -3; j1 <= 4; ++j1) {
 				for (int k1 = -3; k1 <= 4; ++k1) {
 					worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX_COLOR_NORMAL);
-					float f22 = (float) (j1 * 8);
-					float f23 = (float) (k1 * 8);
+					float f22 = j1 * 8;
+					float f23 = k1 * 8;
 					float f24 = f22 - f19;
 					float f25 = f23 - f20;
 
 					if (f3 > -5.0F) {
-						worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 0.0F), (double) (f25 + 8.0F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18)).color(f10, f11, f12, 0.555F)
+						worldrenderer.pos(f24 + 0.0F, f3 + 0.0F, f25 + 8.0F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18).color(f10, f11, f12, 0.555F)
 								.normal(0.0F, -1.0F, 0.0F).endVertex();
-						worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 0.0F), (double) (f25 + 8.0F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18)).color(f10, f11, f12, 0.555F)
+						worldrenderer.pos(f24 + 8.0F, f3 + 0.0F, f25 + 8.0F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18).color(f10, f11, f12, 0.555F)
 								.normal(0.0F, -1.0F, 0.0F).endVertex();
-						worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 0.0F), (double) (f25 + 0.0F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18)).color(f10, f11, f12, 0.555F)
+						worldrenderer.pos(f24 + 8.0F, f3 + 0.0F, f25 + 0.0F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18).color(f10, f11, f12, 0.555F)
 								.normal(0.0F, -1.0F, 0.0F).endVertex();
-						worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 0.0F), (double) (f25 + 0.0F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18)).color(f10, f11, f12, 0.555F)
+						worldrenderer.pos(f24 + 0.0F, f3 + 0.0F, f25 + 0.0F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18).color(f10, f11, f12, 0.555F)
 								.normal(0.0F, -1.0F, 0.0F).endVertex();
 					}
 
 					if (f3 <= 5.0F) {
-						worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 7.0F - 9.765625E-4F), (double) (f25 + 8.0F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18)).color(f4, f5, f6, 0.555F)
+						worldrenderer.pos(f24 + 0.0F, f3 + 7.0F - 9.765625E-4F, f25 + 8.0F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18).color(f4, f5, f6, 0.555F)
 								.normal(0.0F, 1.0F, 0.0F).endVertex();
-						worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 7.0F - 9.765625E-4F), (double) (f25 + 8.0F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18)).color(f4, f5, f6, 0.555F)
+						worldrenderer.pos(f24 + 8.0F, f3 + 7.0F - 9.765625E-4F, f25 + 8.0F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18).color(f4, f5, f6, 0.555F)
 								.normal(0.0F, 1.0F, 0.0F).endVertex();
-						worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 7.0F - 9.765625E-4F), (double) (f25 + 0.0F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18)).color(f4, f5, f6, 0.555F)
+						worldrenderer.pos(f24 + 8.0F, f3 + 7.0F - 9.765625E-4F, f25 + 0.0F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18).color(f4, f5, f6, 0.555F)
 								.normal(0.0F, 1.0F, 0.0F).endVertex();
-						worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 7.0F - 9.765625E-4F), (double) (f25 + 0.0F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18)).color(f4, f5, f6, 0.555F)
+						worldrenderer.pos(f24 + 0.0F, f3 + 7.0F - 9.765625E-4F, f25 + 0.0F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18).color(f4, f5, f6, 0.555F)
 								.normal(0.0F, 1.0F, 0.0F).endVertex();
 					}
 
 					if (j1 > -1) {
 						for (int l1 = 0; l1 < 8; ++l1) {
-							worldrenderer.pos((double) (f24 + (float) l1 + 0.0F), (double) (f3 + 0.0F), (double) (f25 + 8.0F)).tex((double) ((f22 + (float) l1 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + l1 + 0.0F, f3 + 0.0F, f25 + 8.0F).tex((f22 + l1 + 0.5F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(-1.0F, 0.0F, 0.0F).endVertex();
-							worldrenderer.pos((double) (f24 + (float) l1 + 0.0F), (double) (f3 + 7.0F), (double) (f25 + 8.0F)).tex((double) ((f22 + (float) l1 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + l1 + 0.0F, f3 + 7.0F, f25 + 8.0F).tex((f22 + l1 + 0.5F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(-1.0F, 0.0F, 0.0F).endVertex();
-							worldrenderer.pos((double) (f24 + (float) l1 + 0.0F), (double) (f3 + 7.0F), (double) (f25 + 0.0F)).tex((double) ((f22 + (float) l1 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + l1 + 0.0F, f3 + 7.0F, f25 + 0.0F).tex((f22 + l1 + 0.5F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(-1.0F, 0.0F, 0.0F).endVertex();
-							worldrenderer.pos((double) (f24 + (float) l1 + 0.0F), (double) (f3 + 0.0F), (double) (f25 + 0.0F)).tex((double) ((f22 + (float) l1 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + l1 + 0.0F, f3 + 0.0F, f25 + 0.0F).tex((f22 + l1 + 0.5F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(-1.0F, 0.0F, 0.0F).endVertex();
 						}
 					}
 
 					if (j1 <= 1) {
 						for (int i2 = 0; i2 < 8; ++i2) {
-							worldrenderer.pos((double) (f24 + (float) i2 + 1.0F - 9.765625E-4F), (double) (f3 + 0.0F), (double) (f25 + 8.0F)).tex((double) ((f22 + (float) i2 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + i2 + 1.0F - 9.765625E-4F, f3 + 0.0F, f25 + 8.0F).tex((f22 + i2 + 0.5F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(1.0F, 0.0F, 0.0F).endVertex();
-							worldrenderer.pos((double) (f24 + (float) i2 + 1.0F - 9.765625E-4F), (double) (f3 + 7.0F), (double) (f25 + 8.0F)).tex((double) ((f22 + (float) i2 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 8.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + i2 + 1.0F - 9.765625E-4F, f3 + 7.0F, f25 + 8.0F).tex((f22 + i2 + 0.5F) * 0.00390625F + f17, (f23 + 8.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(1.0F, 0.0F, 0.0F).endVertex();
-							worldrenderer.pos((double) (f24 + (float) i2 + 1.0F - 9.765625E-4F), (double) (f3 + 7.0F), (double) (f25 + 0.0F)).tex((double) ((f22 + (float) i2 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + i2 + 1.0F - 9.765625E-4F, f3 + 7.0F, f25 + 0.0F).tex((f22 + i2 + 0.5F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(1.0F, 0.0F, 0.0F).endVertex();
-							worldrenderer.pos((double) (f24 + (float) i2 + 1.0F - 9.765625E-4F), (double) (f3 + 0.0F), (double) (f25 + 0.0F)).tex((double) ((f22 + (float) i2 + 0.5F) * 0.00390625F + f17), (double) ((f23 + 0.0F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + i2 + 1.0F - 9.765625E-4F, f3 + 0.0F, f25 + 0.0F).tex((f22 + i2 + 0.5F) * 0.00390625F + f17, (f23 + 0.0F) * 0.00390625F + f18)
 									.color(f26, f27, f28, 0.555F).normal(1.0F, 0.0F, 0.0F).endVertex();
 						}
 					}
 
 					if (k1 > -1) {
 						for (int j2 = 0; j2 < 8; ++j2) {
-							worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 7.0F), (double) (f25 + (float) j2 + 0.0F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + (float) j2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 0.0F, f3 + 7.0F, f25 + j2 + 0.0F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + j2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, -1.0F).endVertex();
-							worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 7.0F), (double) (f25 + (float) j2 + 0.0F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + (float) j2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 8.0F, f3 + 7.0F, f25 + j2 + 0.0F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + j2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, -1.0F).endVertex();
-							worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 0.0F), (double) (f25 + (float) j2 + 0.0F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + (float) j2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 8.0F, f3 + 0.0F, f25 + j2 + 0.0F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + j2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, -1.0F).endVertex();
-							worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 0.0F), (double) (f25 + (float) j2 + 0.0F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + (float) j2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 0.0F, f3 + 0.0F, f25 + j2 + 0.0F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + j2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, -1.0F).endVertex();
 						}
 					}
 
 					if (k1 <= 1) {
 						for (int k2 = 0; k2 < 8; ++k2) {
-							worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 7.0F), (double) (f25 + (float) k2 + 1.0F - 9.765625E-4F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + (float) k2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 0.0F, f3 + 7.0F, f25 + k2 + 1.0F - 9.765625E-4F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + k2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, 1.0F).endVertex();
-							worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 7.0F), (double) (f25 + (float) k2 + 1.0F - 9.765625E-4F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + (float) k2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 8.0F, f3 + 7.0F, f25 + k2 + 1.0F - 9.765625E-4F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + k2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, 1.0F).endVertex();
-							worldrenderer.pos((double) (f24 + 8.0F), (double) (f3 + 0.0F), (double) (f25 + (float) k2 + 1.0F - 9.765625E-4F)).tex((double) ((f22 + 8.0F) * 0.00390625F + f17), (double) ((f23 + (float) k2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 8.0F, f3 + 0.0F, f25 + k2 + 1.0F - 9.765625E-4F).tex((f22 + 8.0F) * 0.00390625F + f17, (f23 + k2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, 1.0F).endVertex();
-							worldrenderer.pos((double) (f24 + 0.0F), (double) (f3 + 0.0F), (double) (f25 + (float) k2 + 1.0F - 9.765625E-4F)).tex((double) ((f22 + 0.0F) * 0.00390625F + f17), (double) ((f23 + (float) k2 + 0.5F) * 0.00390625F + f18))
+							worldrenderer.pos(f24 + 0.0F, f3 + 0.0F, f25 + k2 + 1.0F - 9.765625E-4F).tex((f22 + 0.0F) * 0.00390625F + f17, (f23 + k2 + 0.5F) * 0.00390625F + f18)
 									.color(f13, f14, f15, 0.555F).normal(0.0F, 0.0F, 1.0F).endVertex();
 						}
 					}
